fix: change incident treshhold column format

+++ server/apps/core/incident_types.py
@@ -5,7 +5,7 @@ from server.settings.components.common import BASE_DIR, MODELS_DIR
 
 class IncidentType(models.Model):
     model_path = models.CharField(max_length=100, null=True)
-    treshold = models.PositiveIntegerField('Treshold для модели', default=1)
+    treshold = models.FloatField('Treshold для модели', default=1)
     description = models.CharField(
         'Название модели', max_length=128, null=True, blank=True)
     chat_gpt_prompt = models.TextField('Chat-GPR промпт', null=True, blank=True)
